changed ensemble to an ocean one. Print statements removed. Typos cleared. conftest.py with pytest fixtures written.
